lessons/lesson.33/zoo/requirements.txt
amqp==5.0.6
asgiref==3.3.4
billiard==3.6.4.0
celery==5.1.1
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.4
django-debug-toolbar==3.2.1
factory-boy==3.2.0
Faker==8.9.1
faker-music==0.4
kombu==5.1.0
mixer==7.1.2
Pillow==8.2.0
prompt-toolkit==3.0.19
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
vine==5.0.0
wcwidth==0.2.5
